test: keep event bridge regression within package budget
